resolve homepage in gatsby-node
